+++ uninstall.php
@@ -1,13 +1,13 @@
 <?php // ------------------------------------------------------------------------------------------------------------------------ //
 
-require_once('usi-debug-enable.php');
-require_once('usi-dbs-mysqli.php');
-require_once('usi-settings-uninstall.php');
+require_once('usi-library/usi-debug-enable.php');
+require_once('usi-library/usi-dbs-mysqli.php');
+require_once('usi-settings/usi-settings-uninstall.php');
 require_once('usi-page-solutions.php');
 
 final class USI_Page_Solutions_Uninstall {
 
-   const VERSION = '1.0.0 (2018-01-07)';
+   const VERSION = '1.1.0 (2018-01-13)';
 
    private function __construct() {
    } // __construct();