+++ usi-page-solutions-virtual.php
@@ -4,7 +4,7 @@ defined('ABSPATH') or die('Accesss not allowed.');
 
 class USI_Page_Solutions_Virtual {
 
-   const VERSION = '1.0.1 (2018-01-07)';
+   const VERSION = '1.1.0 (2018-01-13)';
 
    private $options = null;
    private $option_name = null;