+++ usi-page-solutions.php
@@ -5,8 +5,8 @@ defined('ABSPATH') or die('Accesss not allowed.');
 /*
 Plugin Name: Page-Solutions
 Plugin URI: https://github.com/jaschwanda/page-solutions
-Description: The Page-Solutions plugin extends the WordPress widget system by enabling the creation of virtual widget collections that can be displayed in enhanced widget areas on a page by page basis. It also provides page caching functionality. The Page-Solutions plugin is developed and maintained by Universal Solutions. 
-Version: 1.0.0 (2018-01-07)
+Description: The Page-Solutions plugin provides custom CSS and JavaScript modifications, virtual widget mapping and page caching functionality on a page by page basis. It is a thin plugin that enhances WordPress functionality for page-intensive and non-blog applications. The Page-Solutions plugin is developed and maintained by Universal Solutions.
+Version: 1.1.0 (2018-01-13)
 Author: Jim Schwanda
 Author URI: http://www.usi2solve.com/leader
 Text Domain: usi-page-solutions
@@ -17,7 +17,7 @@ require_once('usi-settings/usi-settings.php');
 
 final class USI_Page_Solutions {
 
-   const VERSION = '1.0.1 (2018-01-07)';
+   const VERSION = '1.1.0 (2018-01-13)';
 
    const DEBUG_OFF   = 0x00;
    const DEBUG_HTML  = 0x01;